2020. 09. 24 <2> Stage Clear 시 Clear Image가 활성화되도록 표기(FileManager의 StageClear로 관리) Gold 스크립트 및 3D모델링 적용해서 생동감 있는 코인 모션 연출

+++ 2020_2_Project/Assets/Scripts/Manager/FileManager.cs
@@ -13,7 +13,7 @@ using System.Text;
 
 public class FileManager : MonoBehaviour
 {
-    public static Dictionary<string, float> playerInfo = new Dictionary<string, float>();
+    public static Dictionary<string, int> playerInfo = new Dictionary<string, int>();
     public static Dictionary<string, bool> soundSetting = new Dictionary<string, bool>();
     public static Dictionary<string, bool> stageClear = new Dictionary<string, bool>();
